+++ tests/Pim/Product/ProductUpsertDataTest.php
@@ -78,6 +78,19 @@ final class ProductUpsertDataTest extends TestCase
         );
     }
 
+    public function testQuotesData(): void
+    {
+        $productUpsertData = $this->minimumValidData();
+        $productUpsertData
+            ->setQuoteRequestsMinQuantity(10)
+            ->setIsExclusiveToQuoteRequests(true);
+        $productUpsertData->validate();
+
+        $returnData = $productUpsertData->toArray();
+        static::assertEquals(10, $returnData['quote_requests_min_quantity']);
+        static::assertEquals(true, $returnData['is_exclusive_to_quote_requests']);
+    }
+
     private function minimumValidData(): ProductUpsertData
     {
         return (new CreateProductCommand())